requirements.txt
dnspython==2.6.1
requests==2.32.3
pytest==8.3.3
pydantic==2.9.2
pydantic-settings==2.6.1

